[fix] bug修复

+++ EMS/Model/SmartMeterModel.cs
@@ -9,20 +9,20 @@ namespace EMS.Model
 {
     public class SmartMeterModel
     {
-        public int Voltage_A { get; set; }
-        public int Voltage_B { get; set; }
-        public int Voltage_C { get; set; }
-        public int Current_A {  get; set; }
-        public int Current_B {  get; set; }
-        public int Current_C {  get; set; }
-        public int ActivePower_A {  get; set; }
-        public int ActivePower_B {  get; set; }
-        public int ActivePower_C {  get; set; }
-        public int ActivePower_Total {  get; set; }
-        public int ReactivePower_A { get; set; }
-        public int ReactivePower_B { get; set; }
-        public int ReactivePower_C { get; set; }
-        public int ReactivePower_Total { get; set; }
+        public double Voltage_A { get; set; }
+        public double Voltage_B { get; set; }
+        public double Voltage_C { get; set; }
+        public double Current_A {  get; set; }
+        public double Current_B {  get; set; }
+        public double Current_C {  get; set; }
+        public double ActivePower_A {  get; set; }
+        public double ActivePower_B {  get; set; }
+        public double ActivePower_C {  get; set; }
+        public double ActivePower_Total {  get; set; }
+        public double ReactivePower_A { get; set; }
+        public double ReactivePower_B { get; set; }
+        public double ReactivePower_C { get; set; }
+        public double ReactivePower_Total { get; set; }
         public string SmartMeterNumber {  get; set; }
         public object Clone()
         {
